workflows/prompt-chain-example/pyproject.toml
[project]
name = "prompt-chain-example"
version = "0.1.0"
description = "Example of a simple sequential prompt chain for legal document summarization and review."
authors = [
    {name = "Malte Hedderich",email = "github@hedderich.pro"}
]
readme = "README.md"

packages = [{ include = "prompt_chain_example" }]


requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "streamlit (>=1.42.0,<2.0.0)",
    "google-genai (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
black = "^25.1.0"
flake8 = "^7.1.1"
watchdog = "^6.0.0"
google-api-python-client-stubs = "^1.29.0"

[tool.mypy]
plugins = ["pydantic.mypy"]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
